Starting rewrite in C
Why? Because I want to learn how to write programs in C. So I chose
this program, because it appears simple enough to me, and it will
help me to learn how to write a complete C program.

While I'm at it, I'm trying to expand the program to perform the
other features documented in the TODO.